fix(665): Clean up
